# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Clinic API
  version: 1.0.0
  description: 'API contract for the Dental Clinic MVP. Authenticate with `POST /api/auth/login/`
    and send `Authorization: Token <token>`.'
paths:
...
  /api/visits/{id}/follow-ups/:
    post:
      operationId: visits_follow_ups_create
      description: Book a follow-up appointment with the same doctor (VISIT-006).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowUpRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FollowUpRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FollowUpRequest'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
  /api/visits/{id}/medications/:
    post:
      operationId: visits_medications_create
      description: Prescribe a catalog medication with quantity and duration (MED-002).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitMedicationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitMedicationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitMedicationRequest'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
...
  /api/visits/{id}/procedures/:
    post:
      operationId: visits_procedures_create
      description: Record tooth/procedure information (VISIT-003).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitProcedureRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitProcedureRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitProcedureRequest'
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
...
  /api/visits/{id}/tooth-actions/:
    post:
      operationId: visits_tooth_actions_create
      description: Mark an action on a tooth of this visit's dental chart (CHART-002).
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this visit.
        required: true
      tags:
      - visits
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitToothActionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VisitToothActionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VisitToothActionRequest'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
          description: ''
...
components:
  schemas:
...
    AppointmentStatusEnum:
      enum:
      - SCHEDULED
      - CHECKED_IN
      - IN_VISIT
      - COMPLETED
      - CANCELLED
      type: string
      description: |-
        * `SCHEDULED` - Scheduled
        * `CHECKED_IN` - Waiting for doctor
        * `IN_VISIT` - In visit
        * `COMPLETED` - Completed
        * `CANCELLED` - Cancelled
...
    DentalActionType:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        code:
          type: string
          description: Optional internal or ADA code.
          maxLength: 20
        color:
          type: string
          description: 'Tooth color on the dental chart, e.g. #2563EB.'
          pattern: ^#[0-9A-Fa-f]{6}$
          maxLength: 7
      required:
      - id
      - name
...
    DoctorSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        full_name:
          type: string
          maxLength: 150
      required:
      - full_name
      - id
...
    FollowUp:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        scheduled_at:
          type: string
          format: date-time
        notes:
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/AppointmentStatusEnum'
          readOnly: true
        status_display:
          type: string
          readOnly: true
      required:
      - id
      - scheduled_at
      - status
      - status_display
    FollowUpRequest:
      type: object
      properties:
        scheduled_at:
          type: string
          format: date-time
        notes:
          type: string
      required:
      - scheduled_at
...
    Medication:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        details:
          type: string
          description: Optional strength/form, e.g. 500 mg capsule.
          maxLength: 200
      required:
      - id
      - name
...
    PatientSummary:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        full_name:
          type: string
          maxLength: 150
        phone:
          type: string
          maxLength: 16
        is_minor:
          type: boolean
          title: Patient is a minor
      required:
      - full_name
      - id
      - phone
...
    Procedure:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        code:
          type: string
          description: Optional internal or ADA code.
          maxLength: 20
      required:
      - id
      - name
...
    Visit:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/VisitStatusEnum'
          readOnly: true
        status_display:
          type: string
          readOnly: true
        patient:
          allOf:
          - $ref: '#/components/schemas/PatientSummary'
          readOnly: true
        doctor:
          allOf:
          - $ref: '#/components/schemas/DoctorSummary'
          readOnly: true
        appointment:
          type: integer
          readOnly: true
        started_at:
          type: string
          format: date-time
          readOnly: true
        completed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        notes:
          type: string
          title: Visit notes
        diagnosis:
          type: string
        treatment:
          type: string
        procedures:
          type: array
          items:
            $ref: '#/components/schemas/VisitProcedure'
          readOnly: true
        tooth_actions:
          type: array
          items:
            $ref: '#/components/schemas/VisitToothAction'
          readOnly: true
        medications:
          type: array
          items:
            $ref: '#/components/schemas/VisitMedication'
          readOnly: true
        follow_ups:
          type: array
          items:
            $ref: '#/components/schemas/FollowUp'
          readOnly: true
        can_edit:
          type: boolean
          readOnly: true
      required:
      - appointment
      - can_edit
      - completed_at
      - doctor
      - follow_ups
      - id
      - medications
      - patient
      - procedures
      - started_at
      - status
      - status_display
      - tooth_actions
    VisitMedication:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        medication:
          allOf:
          - $ref: '#/components/schemas/Medication'
          readOnly: true
        quantity:
          type: string
          maxLength: 100
        duration:
          type: string
          maxLength: 100
      required:
      - duration
      - id
      - medication
      - quantity
    VisitMedicationRequest:
      type: object
      properties:
        medication_id:
          type: integer
          writeOnly: true
        quantity:
          type: string
          minLength: 1
          maxLength: 100
        duration:
          type: string
          minLength: 1
          maxLength: 100
      required:
      - duration
      - medication_id
      - quantity
    VisitProcedure:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        procedure:
          allOf:
          - $ref: '#/components/schemas/Procedure'
          readOnly: true
        tooth:
          type: string
          maxLength: 2
        notes:
          type: string
      required:
      - id
      - procedure
    VisitProcedureRequest:
      type: object
      properties:
        procedure_id:
          type: integer
          writeOnly: true
          nullable: true
          description: Catalog procedure, where applicable (DX-001).
        tooth:
          type: string
          maxLength: 2
        notes:
          type: string
...
    VisitStatusEnum:
      enum:
      - ACTIVE
      - COMPLETED
      type: string
      description: |-
        * `ACTIVE` - Active
        * `COMPLETED` - Completed
    VisitToothAction:
      type: object
      description: One action on one tooth of the visit's dental chart (CHART-002).
      properties:
        id:
          type: integer
          readOnly: true
        tooth:
          type: string
          description: 'FDI notation: 11-48 permanent, 51-85 primary.'
          maxLength: 2
        action_type:
          allOf:
          - $ref: '#/components/schemas/DentalActionType'
          readOnly: true
        notes:
          type: string
          maxLength: 255
      required:
      - action_type
      - id
      - tooth
    VisitToothActionRequest:
      type: object
      description: One action on one tooth of the visit's dental chart (CHART-002).
      properties:
        tooth:
          type: string
          minLength: 1
          description: 'FDI notation: 11-48 permanent, 51-85 primary.'
          maxLength: 2
        action_type_id:
          type: integer
          writeOnly: true
          description: Active dental action type available to the clinic (CHART-001).
        notes:
          type: string
          maxLength: 255
      required:
      - action_type_id
      - tooth
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"